move examples to scripts (#689)
