+++ examples/interactive.py
@@ -14,38 +14,12 @@ Then enter something like:
 as the user input (or in general for the drqa model, enter
 a context followed by '\n' followed by a question all as a single input.)
 """
-from parlai.core.params import ParlaiParser
-from parlai.core.agents import create_agent
-from parlai.core.worlds import create_task
-
+from parlai.scripts.interactive import setup_args, interactive
 import random
 
-def interactive(opt):
-    # Create model and assign it to the specified task
-    opt['task'] = 'parlai.agents.local_human.local_human:LocalHumanAgent'
-
-    agent = create_agent(opt)
-    world = create_task(opt, agent)
-
-    # Show some example dialogs:
-    while True:
-        world.parley()
-        if opt.get('display_examples'):
-            print("---")
-            print(world.display() + "\n~~")
-        if world.epoch_done():
-            print("EPOCH DONE")
-            break
-
-def main():
-    random.seed(42)
-
-    # Get command line arguments
-    parser = ParlaiParser(True, True)
-    parser.add_argument('-d', '--display-examples', type='bool', default=False)
-    opt = parser.parse_args()
-    print(opt)
-    interactive(opt)
 
 if __name__ == '__main__':
-    main()
+    random.seed(42)
+    parser = setup_args()
+    opt = parser.parse_args()
+    interactive(opt)
