+++ examples/profile_train.py
@@ -9,69 +9,10 @@ For example, to make sure that bAbI task 1 (1k exs) loads one can run and to
 see a few of them:
 `python examples/profile.py -t babi:task1k:1 -m seq2seq -e 0.1 --dict-file /tmp/dict`'
 """
-
-import cProfile
-import io
-import pdb
-import pstats
-import torch
-from train_model import setup_args as train_args
-from train_model import TrainLoop
-
-
-def setup_args():
-    parser = train_args()
-    profile = parser.add_argument_group('Profiler Arguments')
-    profile.add_argument('--torch', type='bool', default=False,
-                         help='If true, use the torch profiler. Otherwise use '
-                              'use cProfile.')
-    profile.add_argument('--torch-cuda', type='bool', default=False,
-                         help='If true, use the torch cuda profiler. Otherwise use '
-                              'use cProfile.')
-    profile.add_argument('--debug', type='bool', default=False,
-                         help='If true, enter debugger at end of run.')
-    return parser
-
-
-def main(parser):
-    opt = parser.parse_args()
-
-    if opt['torch'] or opt['torch_cuda']:
-        with torch.autograd.profiler.profile(use_cuda=opt['torch_cuda']) as prof:
-            TrainLoop(parser).train()
-        print(prof.total_average())
-
-        sort_cpu = sorted(prof.key_averages(), key=lambda k: k.cpu_time)
-        sort_cuda = sorted(prof.key_averages(), key=lambda k: k.cuda_time)
-
-        def cpu():
-            for e in sort_cpu:
-                print(e)
-
-        def cuda():
-            for e in sort_cuda:
-                print(e)
-
-        cpu()
-
-        if opt['debug']:
-            print('`cpu()` prints out cpu-sorted list, '
-                  '`cuda()` prints cuda-sorted list')
-
-            pdb.set_trace()
-    else:
-        pr = cProfile.Profile()
-        pr.enable()
-        TrainLoop(parser).train()
-        pr.disable()
-        s = io.StringIO()
-        sortby = 'cumulative'
-        ps = pstats.Stats(pr, stream=s).sort_stats(sortby)
-        ps.print_stats()
-        print(s.getvalue())
-        if opt['debug']:
-            pdb.set_trace()
+from parlai.scripts.profile_train import setup_args, profile
 
 
 if __name__ == '__main__':
-    main(setup_args())
+    parser = setup_args()
+    opt = parser.parse_args()
+    profile(opt)
