+++ projects/convai2/baselines/profilememory/train.py
@@ -8,7 +8,7 @@ These parameters have some variance in their final perplexity, but they were
 used to achieve the pre-trained model.
 """
 
-from examples.train_model import setup_args, TrainLoop
+from parlai.scripts.train_model import setup_args, TrainLoop
 
 
 if __name__ == '__main__':
@@ -40,4 +40,4 @@ if __name__ == '__main__':
         validation_patience=10,
         log_every_n_secs=10,
     )
-    TrainLoop(parser).train()
+    TrainLoop(parser.parse_args()).train()
