+++ projects/convai2/baselines/seq2seq/eval_hits.py
@@ -21,8 +21,8 @@ if __name__ == '__main__':
         rank_candidates=True,
         batchsize=64,
     )
-    opt = parser.parse_args()
+    opt = parser.parse_args(print_args=False)
     opt['model_type'] = 'seq2seq'
     fnames = ['convai2_self_seq2seq_model.tgz', 'dict_convai2_self']
     download_models(opt, fnames, 'convai2')
-    eval_model(parser, printargs=False)
+    eval_model(opt, print_parser=parser)
