+++ projects/personachat/scripts/kvmemnn_eval.py
@@ -1,6 +1,11 @@
+# Copyright (c) 2017-present, Facebook, Inc.
+# All rights reserved.
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree. An additional grant
+# of patent rights can be found in the PATENTS file in the same directory.
 from parlai.core.build_data import download_models
 from parlai.core.params import ParlaiParser
-from examples.eval_model import eval_model
+from parlai.scripts.eval_model import eval_model
 
 '''Evaluate pre-trained model trained for hits@1 metric
 Key-Value Memory Net model trained on personachat using persona 'self'
