+++ projects/personachat/scripts/languagemodel_opensub2018_interactive.py
@@ -1,6 +1,11 @@
+# Copyright (c) 2017-present, Facebook, Inc.
+# All rights reserved.
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree. An additional grant
+# of patent rights can be found in the PATENTS file in the same directory.
 from parlai.core.build_data import download_models
 from parlai.core.params import ParlaiParser
-from examples.interactive import interactive
+from parlai.scripts.interactive import interactive
 from parlai.agents.language_model.language_model import LanguageModelAgent
 
 '''Interact with pre-trained model
